[rsql] Re-organize rsql-jooq test
